@@ -1,30 +1,32 @@
 {
-  "name": "widgets",
-  "version": "1.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "gulp watch",
-    "build": "gulp --production"
-  },
-  "dependencies": {
-    "bootstrap-stylus": "hworld/bootstrap-stylus#master",
-    "hint.css": "2.4.1",
-    "core-js": "2.4.1",
-    "qs": "6.3.0",
-    "superagent": "3.4.1",
-    "vee-validate": "2.0.0-beta.22",
-    "vue": "2.1.10",
-    "vue-class-component": "4.4.0",
-    "vue-property-decorator": "3.4.0",
-    "vuex": "2.1.2",
-    "vuex-class": "0.1.3",
-    "ua-parser-js": "0.7.12",
-    "reflect-metadata": "0.1.9"
-  },
-  "devDependencies": {
-    "@types/qs": "6.2.30",
-    "@types/superagent": "2.0.36",
-    "@types/ua-parser-js": "0.7.30",
-    "game-jolt-frontend-lib": "gamejolt/frontend-lib#master"
-  }
+	"name": "widgets",
+	"version": "1.0.0",
+	"private": true,
+	"scripts": {
+		"dev": "gulp watch",
+		"build": "gulp --production"
+	},
+	"dependencies": {
+		"axios": "0.16.2",
+		"bootstrap-stylus": "hworld/bootstrap-stylus#master",
+		"core-js": "2.5.0",
+		"qs": "6.5.0",
+		"reflect-metadata": "0.1.10",
+		"tslint": "5.6.0",
+		"ua-parser-js": "0.7.14",
+		"v-tooltip": "2.0.0-beta.8",
+		"vee-validate": "2.0.0-rc.9",
+		"vue": "2.4.2",
+		"vue-class-component": "5.0.2",
+		"vue-gettext": "2.0.22",
+		"vue-property-decorator": "5.2.1",
+		"vue-server-renderer": "2.4.2",
+		"vuex": "2.3.1",
+		"vuex-class": "0.2.0"
+	},
+	"devDependencies": {
+		"@types/qs": "6.5.0",
+		"@types/ua-parser-js": "0.7.31",
+		"game-jolt-frontend-lib": "gamejolt/frontend-lib#master"
+	}
 }
