@@ -1,20 +1,22 @@
-import * as Vue from 'vue';
+import Vue from 'vue';
+import { Mutation, Action } from 'vuex-class';
 import { Component } from 'vue-property-decorator';
 import * as View from '!view!./address.html?style=./address.styl';
 
-import { Mutations, Actions, AddressData } from '../../store/index';
+import { AddressData, Store } from '../../store/index';
 import { AppJolticon } from '../../../lib/gj-lib-client/vue/components/jolticon/jolticon';
-import { Geo } from '../../../lib/gj-lib-client/components/geo/geo.service';
+import { Geo, Region, Country } from '../../../lib/gj-lib-client/components/geo/geo.service';
 
 @View
 @Component({
-	name: 'address',
 	components: {
 		AppJolticon,
-	}
+	},
 })
-export class AppAddress extends Vue
-{
+export class AppAddress extends Vue {
+	@Mutation setAddress: Store['setAddress'];
+	@Action checkout: Store['checkout'];
+
 	address = {
 		country: 'us',
 		region: '',
@@ -22,25 +24,27 @@ export class AppAddress extends Vue
 		postcode: '',
 	};
 
-	countries = Geo.getCountries();
-	regions = Geo.getRegions( this.address.country );
+	countries: Country[] = [];
+	regions: Region[] | null = [];
 
-	countryChanged()
-	{
-		this.regions = Geo.getRegions( this.address.country );
+	created() {
+		this.countries = Geo.getCountries();
+		this.countryChanged();
+	}
 
-		if ( this.regions ) {
+	countryChanged() {
+		this.regions = Geo.getRegions(this.address.country) || null;
+
+		if (this.regions) {
 			this.address.region = this.regions[0].code;
-		}
-		else {
+		} else {
 			this.address.region = '';
 		}
 	}
 
-	submit()
-	{
-		const addressData = Object.assign( new AddressData(), this.address );
-		this.$store.commit( Mutations.setAddress, addressData );
-		this.$store.dispatch( Actions.checkout );
+	submit() {
+		const addressData = Object.assign(new AddressData(), this.address);
+		this.setAddress(addressData);
+		this.checkout();
 	}
 }
